Implemented Falling and Respawning in Pits
Player can now detect if he's falling in a pit, if he is, he changes the state to "Respawn State" which just makes the respawn happen and could decrease the players health. The "Respawn State" then transitions back to the move state (or later maybe dead state, if he lost too much health).